+++ ui/src/main/java/de/ipb_halle/lbac/project/ProjectService.java
@@ -36,18 +36,17 @@ import de.ipb_halle.lbac.search.lang.SqlBuilder;
 import de.ipb_halle.lbac.search.lang.Value;
 import de.ipb_halle.lbac.service.NodeService;
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import javax.annotation.PostConstruct;
-import javax.ejb.Stateless;
-import javax.inject.Inject;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
+import jakarta.annotation.PostConstruct;
+import jakarta.ejb.Stateless;
+import jakarta.inject.Inject;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import jakarta.persistence.Query;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -179,7 +178,7 @@ public class ProjectService implements Serializable {
     }
 
     public boolean isProjectNameAvailable(String name) {
-        BigInteger i = (BigInteger) this.em.createNativeQuery(
+        Long i = (Long) this.em.createNativeQuery(
                 SQL_GET_NAME_AVAILABLE)
                 .setParameter("name", name)
                 .getResultList().get(0);
